+++ src/clj/x/server_installer/db_installer.clj
@@ -28,5 +28,6 @@
   :x.server-installer/install-db!
   ; WARNING! NON-PUBLIC! DO NOT USE!
   (fn [{:keys [db]} _]
-      (println details/app-name "module-installed: db")
-      {:db (r engine/->module-installed db :db)}))
+      (println details/app-name "installing module: db")
+      {:db       (r engine/->module-installed db :db)
+       :dispatch [:mongo-db/connect!]}))