+++ src/clj/x/server_installer/user_installer.clj
@@ -29,9 +29,10 @@
   :x.server-installer/install-user!
   ; WARNING! NON-PUBLIC! DO NOT USE!
   (fn [{:keys [db]} _]
-      (println details/app-name "module-installed: user")
-      {:db (r engine/->module-installed db :user)}))
-       ;:dispatch [:x.server-user/add-user! :anonymous {}]}))
+      (println details/app-name "installing module: user")
+      {:db (r engine/->module-installed db :user)
+       :dispatch [:x.server-user/add-user! {:email-address "demo@monotech.hu"
+                                            :password      "mono"}]}))