+++ src/cljs/x/app_core/connection_handler.cljs
@@ -5,8 +5,8 @@
 ; Author: bithandshake
 ; Created: 2021.05.20
 ; Description:
-; Version: v0.2.8
-; Compatibility: x4.2.5
+; Version: v0.3.4
+; Compatibility: x4.4.2
 
 
 
@@ -14,8 +14,8 @@
 ;; ----------------------------------------------------------------------------
 
 (ns x.app-core.connection-handler
-    (:require [x.app-core.api     :as a :refer [r]]
-              [mid-fruits.vector  :as vector]
+    (:require [mid-fruits.vector            :as vector]
+              [x.app-core.event-handler     :as event-handler :refer [r]]
               [x.app-core.lifecycle-handler :as lifecycle-handler]))
 
 
@@ -23,13 +23,13 @@
 ;; ----------------------------------------------------------------------------
 ;; ----------------------------------------------------------------------------
 
-(a/reg-event-fx
+(event-handler/reg-event-fx
   :x.app-core/connect-app!
   ; WARNING! NON-PUBLIC! DO NOT USE!
   (fn [{:keys [db]} _]
       {:dispatch-n (r lifecycle-handler/get-period-events db :on-browser-online)}))
 
-(a/reg-event-fx
+(event-handler/reg-event-fx
   :x.app-core/disconnect-app!
   ; WARNING! NON-PUBLIC! DO NOT USE!
   (fn [{:keys [db]} _]