+++ CHANGES.md
@@ -1,8 +1,16 @@
 
+# x4.4.2
+
+- NEW COMPONENT
+  x.app-components.api/infinite-loader
+
+
+
 # x4.4.1
 
 - BREAKING CHANGE
   project-details.edn -> x.project-details.edn
+
 
 
 # x4.4.0